# Conflicts:
#	src/cdrom/cdrom_image.c
#	src/cdrom/cdrom_image_backend.c
#	src/include/86box/cdrom.h
#	src/include/86box/cdrom_image_backend.h
#	src/include/86box/resource.h
#	src/qt/qt_mediamenu.cpp
#	src/qt/qt_mediamenu.hpp
#	src/unix/unix.c
#	src/win/win.c
#	src/win/win_devconf.c
#	src/win/win_dialog.c
#	src/win/win_joystick_rawinput.c
#	src/win/win_media_menu.c
#	src/win/win_mouse.c
#	src/win/win_settings.c
#	src/win/win_ui.c

+++ src/include/86box/snd_sb_dsp.h
@@ -42,7 +42,7 @@ typedef struct sb_dsp_t {
 
     int sb_irqnum;
     void (*irq_update)(void *priv, int set),
-         *irq_priv;
+        *irq_priv;
 
     uint8_t sbe2;
     int     sbe2count;
@@ -83,7 +83,7 @@ typedef struct sb_dsp_t {
 
     pc_timer_t output_timer, input_timer;
 
-    uint64_t sblatcho, sblatchi;
+    double sblatcho, sblatchi;
 
     uint16_t sb_addr;
